Cache Bundled gems, when we can.
By copying just the Gemfile and Gemfile.lock, and then `RUN`-ing the
`bundle install` we can leverage Docker's ability to cache prior runs.
If the Gemfile/Gemfile.lock are later updated, the cache is blown and
all following Docker instructions must be re-done.

We then `COPY` in the actual source for the site, blowing the cache for
all steps which follow it any time a source file has changed.